ai-orchestrator/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# LangChain and AI
langchain==0.1.0
langchain-community==0.0.20
langchain-openai==0.0.2
openai>=1.6.1,<2.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Redis removed - not needed for current implementation

# Logging and monitoring
structlog==23.2.0
python-json-logger==2.0.7

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0

# JWT Authentication
PyJWT==2.8.0

# HTTP client
httpx==0.25.2

# Utilities
python-dotenv==1.0.0

# Vector operations
numpy==1.24.3

# Image processing
Pillow==10.0.1

# Additional AI tools
tiktoken>=0.5.2,<0.6.0

requests>=2.31